修改zk-Latency avg为小数时的数值转换异常问题

+++ km-common/src/main/java/com/xiaojukeji/know/streaming/km/common/bean/entity/zookeeper/fourletterword/ServerCmdData.java
@@ -17,7 +17,7 @@ import lombok.Data;
 @Data
 public class ServerCmdData extends BaseFourLetterWordCmdData {
     private String zkVersion;
-    private Long zkAvgLatency;
+    private Float zkAvgLatency;
     private Long zkMaxLatency;
     private Long zkMinLatency;
     private Long zkPacketsReceived;
